Support auto configuration url proxy
Use the system proxy information from Mono to determine the proxy
information for a proxy that uses a proxy .pac file. The proxy
information returned from CFNetwork.GetProxiesForUri does not
have the proxy host and port information to be able to get the
credentials. Neither does the CFNetwork.GetDefaultProxy method.